RDM-10924 refactor after upstream branch merger
   [RDM-10924](https://tools.hmcts.net/jira/browse/RDM-10924)

Refactor RDM-10924 after merger with attribute-based-access-control
 branch.  Sonar issues and enum changes.
