+++ src/main/java/uk/gov/hmcts/ccd/domain/service/casedataaccesscontrol/FilteredRoleAssignments.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.ccd.domain.service.casedataaccesscontrol;
 
-import uk.gov.hmcts.ccd.domain.model.casedataaccesscontrol.GrantType;
 import uk.gov.hmcts.ccd.domain.model.casedataaccesscontrol.RoleAssignment;
+import uk.gov.hmcts.ccd.domain.model.casedataaccesscontrol.enums.GrantType;
 
 import java.util.ArrayList;
 import java.util.List;
