+++ src/test/java/uk/gov/hmcts/ccd/domain/service/casedataaccesscontrol/FilteredRoleAssignmentsTest.java
@@ -2,8 +2,8 @@ package uk.gov.hmcts.ccd.domain.service.casedataaccesscontrol;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import uk.gov.hmcts.ccd.domain.model.casedataaccesscontrol.GrantType;
 import uk.gov.hmcts.ccd.domain.model.casedataaccesscontrol.RoleAssignment;
+import uk.gov.hmcts.ccd.domain.model.casedataaccesscontrol.enums.GrantType;
 
 import java.util.HashMap;
 import java.util.List;
